requirements.txt
fastapi
Jinja2
python-multipart
uvicorn
deta
python-dateutil